+++ csvtk/cmd/mutate2.go
@@ -292,7 +292,7 @@ Custom functions:
 			var colnames2fileds map[string][]int // column name -> []field
 			var colnamesMap map[string]*regexp.Regexp
 			var fieldsUniq []int
-			var hasHeaderRow bool
+			var selectWithColnames bool
 			var record2 []string // for output
 			keys := make([]string, 0, 8)
 
@@ -305,6 +305,8 @@ Custom functions:
 				if checkFirstLine {
 					checkFirstLine = false
 
+					selectWithColnames = record.SelectWithColnames
+
 					parameters = make(map[string]string, len(record.All))
 					parameters2 = make(map[string]interface{}, len(record.All))
 					parameters2["shenweiNULL"] = nil
@@ -312,8 +314,6 @@ Custom functions:
 					fieldsUniq = UniqInts(record.Fields)
 
 					if !config.NoHeaderRow || record.IsHeaderRow { // do not replace head line
-						hasHeaderRow = true
-
 						colnames2fileds = make(map[string][]int, len(record.All))
 
 						colnamesMap = make(map[string]*regexp.Regexp, len(record.All))
@@ -332,8 +332,8 @@ Custom functions:
 					}
 				}
 
-				// prepaire parameters
-				if !hasHeaderRow {
+				// prepare parameters
+				if !selectWithColnames {
 					for _, fieldTmp = range fieldsUniq {
 						value = record.All[fieldTmp-1]
 						col = strconv.Itoa(fieldTmp)