+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- [csvtk v0.27.1](https://github.com/shenwei356/csvtk/releases/tag/v0.27.1)
+[![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.27.1/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.27.1)
+  - `csvtk filter2/mutate2`:
+      - fix the bug when selecting with field numbers. [#242](https://github.com/shenwei356/csvtk/issues/242)
 - [csvtk v0.27.0](https://github.com/shenwei356/csvtk/releases/tag/v0.27.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.27.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.27.0)
   - `csvtk`: